Feature navigate_to_goal action server (#24)
* Add goal_pose tool

* Remove build warning

* Feature NavigateToGoal action

* Feature navigate_to_goal action server

+++ ike_controller/src/ike_controller.cpp
@@ -113,7 +113,7 @@ std::pair<std::vector<double>, std::vector<double>> IkeController::optimize(
   const std::tuple<double, double, double> & robot_pose,
   const std::pair<std::vector<double>, std::vector<double>> & path)
 {
-  if (path.first.size() < predictive_horizon_num_) {
+  if (path.first.size() < static_cast<long unsigned int>(predictive_horizon_num_)) {
     predictive_horizon_num_ = path.first.size();
   }
